Create utils class and address review comments
